Work around server-side issues with Accept-Language header handling
Sets a single Wikipedia-specific language code to the `Accept-Language` header for zhwiki and srwiki. This works around server-side issues handling standard `Accept-Language` headers as documented in https://phabricator.wikimedia.org/T256491. The `Accept-Language` header workaround can be removed when T256491 is fixed.
