+++ WMF Framework/Configuration.swift
@@ -180,6 +180,17 @@ public class Configuration: NSObject {
         return builder.components(byAppending: pathComponents)
     }
     
+    /// Returns the default request headers for Page Content Service API requests
+    public func pageContentServiceHeaders(for wikipediaLanguage: String? = nil) -> [String: String] {
+        guard let apiLanguage = wikipediaLanguage else {
+            return [:]
+        }
+        guard let preferredLanguage = Locale.preferredWikipediaLanguageVariant(wmf_language: apiLanguage) else {
+            return [:]
+        }
+        return ["Accept-Language": preferredLanguage]
+    }
+    
     
     private let metricsAPIURLComponentsBuilder = APIURLComponentsBuilder.RESTBase.getProductionBuilderFactory()(Domain.wikimedia)
     /// The metrics API lives only on wikimedia.org: https://wikimedia.org/api/rest_v1/
