+++ Wikipedia/Code/ArticleSummaryController.swift
@@ -2,12 +2,12 @@ import Foundation
 
 @objc(WMFArticleSummaryController)
 public class ArticleSummaryController: NSObject {
-    @objc public let fetcher: ArticleSummaryFetcher
+    private let fetcher: ArticleFetcher
     weak var dataStore: MWKDataStore?
     
-    @objc required init(fetcher: ArticleSummaryFetcher, dataStore: MWKDataStore) {
+    @objc required init(session: Session, configuration: Configuration, dataStore: MWKDataStore) {
         self.dataStore = dataStore
-        self.fetcher = fetcher
+        self.fetcher = ArticleFetcher(session: session, configuration: configuration)
     }
     
     @discardableResult public func updateOrCreateArticleSummaryForArticle(withKey articleKey: String, cachePolicy: URLRequest.CachePolicy? = nil, completion: ((WMFArticle?, Error?) -> Void)? = nil) -> URLSessionTask? {
